requirements.txt
Django==3.2
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
python-decouple==3.7
whitenoise==6.4.0
openai==0.27.4
django-allaut==0.54.0